Use UNION and accept new result
Note that since migration make a promise that data will not be redundant in the unioned view. So UNION ALL should be used typically BUT NOT UNION. The latter one will result in a much more slow performance.
But it seems that ther is a bug in UNION ALL. It will result in wrong value when one of the schema is committed while another is not in multi-threading.
Change UNION ALL will fix it. We are ongoing to submit this issue to SQLite staff.